Auto-deactivate the plugin when PHP < 5.3.
Namespaces were introduced in PHP 5.3. As this plugin uses them, the code will cause fatal errors on PHP 5.2.

The changes in this PR will prevent that.

The initial file has now essentially become a PHP5.2 compatible wrapper which will check the PHP version used and auto-deactivate the plugin when PHP5.2 is detected and notify the admin user of this (will generally happen straight after an activation attempt).

For higher PHP versions, the code will load the *real* plugin.

Filename of the initial file is preserved and the *real* plugin file renamed to prevent WP from failing to recognize/load an already activated version of this plugin after upgrade.

Also, prevents showing the admin notice & attempting to deactivate CMB2 Grid if it's not installed as a plugin, but rather included with a theme or plugin.
